Update vscode.mdx
Setting the dart.flutterSdkPath to ".fvm/versions/stable" did not work for me. Because with that it looks into the project folder and not in the home folder of the user. Where the .fvm folder is located. So only by referencing to the home directory with "~/.fvm/versions/stable" worked as it should.